Updated build script

+++ build/Build-Release.ps1
@@ -12,10 +12,21 @@ if($null -ne $pnpframework_response)
 		$pnpframework_hash = $pnpframework_response.commit.sha
 	}
 }
-Write-host "Latest PnP PowerShell Commit hash $pnppowershell_hash" -ForegroundColor Yellow
-Write-Host "Stored PnP PowerShell Commit hash: $existing_pnppowershell_hash" -ForegroundColor Yellow
-Write-host "Latest PnP Framework Commit hash $pnpframework_hash" -ForegroundColor Yellow
-Write-Host "Stored PnP Framework Commit hash: $existing_pnpframework_hash" -ForegroundColor Yellow
+
+$existing_pnpcoresdk_hash = Get-Content ./pnpcoresdk_hash.txt -Raw -ErrorAction SilentlyContinue
+$pnpcoresdk_response = Invoke-RestMethod -Method Get -Uri "$($env:GITHUB_API_URL)/repos/pnp/pnpcore/branches/dev" -SkipHttpErrorCheck
+if($null -ne $pnpcoresdk_response)
+{
+	if($null -ne $pnpcoresdk_response.commit)
+	{
+		$pnpcoresdk_hash = $pnpcoresdk_response.commit.sha
+	}
+}
+
+#Write-host "Latest PnP PowerShell Commit hash $pnppowershell_hash" -ForegroundColor Yellow
+#Write-Host "Stored PnP PowerShell Commit hash: $existing_pnppowershell_hash" -ForegroundColor Yellow
+#Write-host "Latest PnP Framework Commit hash $pnpframework_hash" -ForegroundColor Yellow
+#Write-Host "Stored PnP Framework Commit hash: $existing_pnpframework_hash" -ForegroundColor Yellow
 
 if ($existing_pnppowershell_hash -ne $pnppowershell_hash)
 {
@@ -24,10 +35,17 @@ if ($existing_pnppowershell_hash -ne $pnppowershell_hash)
 	$runPublish = $true
 }
 
-if($existing_pnpframework_hash -ne $pnpframework_hash)
+if($runPublish -eq $false -and $existing_pnpframework_hash -ne $pnpframework_hash)
 {
 	Write-Host "PnP Framework is newer"
 	Set-Content ./pnpframework_hash.txt -Value $pnpframework_hash -NoNewline -Force
+	$runPublish = $true
+}
+
+if($runPublic -eq $false -and $existing_pnpcoresdk_hash -ne $pnpcoresdk_hash)
+{
+	Write-Host "PnP Core SDK is newer"
+	Set-Content ./pnpcoresdk_hash.txt -Value $pnpcoresdk_hash -NoNewLine -Force
 	$runPublish = $true
 }
